misconception_stem_rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.3
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-dotenv==1.0.0
pymupdf4llm==0.0.12
chromadb==0.4.22
sentence-transformers==2.3.1
openai==1.6.1
redis==5.0.1
httpx==0.25.2
aiofiles==23.2.1
numpy<2.0.0,>=1.24.0
nltk>=3.8
spacy>=3.7.0
langchain>=0.1.0
faiss-cpu>=1.7.4
networkx>=3.0
keybert>=0.8.0
scikit-learn>=1.3.0
email-validator==2.1.0.post1
pytest==7.4.3
pytest-asyncio==0.21.1
